requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Config parsing
pyyaml>=6.0

# Scientific computing
numpy>=1.24.0

# Image I/O (PIL backend for torchvision)
Pillow>=9.0.0

# Model complexity metrics
ptflops>=0.7.0

# Additional pretrained models (optional – expands backbone choices)
timm>=0.9.0
